style: change components root
