+++ src/components/Detail/Contents/Information/BasicInformation/BasicInformation.module.scss
