+++ src/components/Detail/Contents/Information/BasicInformation/BasicInformation.tsx
