+++ src/components/Detail/Contents/Information/BasicInformation/Map/Map.tsx
