+++ src/components/Detail/Contents/Information/Others/OtherCard/OtherCard.module.scss
