+++ src/components/Detail/Contents/Information/Others/OtherCard/OtherCard.tsx
