+++ src/components/Detail/Contents/Information/Others/Others.module.scss
