+++ src/components/Detail/Main/Main.module.scss
