+++ src/components/Detail/Main/Main.tsx
