+++ src/components/Detail/Main/Swiper/Swiper.module.scss
