+++ src/components/Detail/Main/Title/Title.module.scss
