+++ src/components/Detail/Navigation/Navigation.module.scss
