+++ src/components/Detail/Navigation/Navigation.tsx
